1. Add the Action with "redux-thunk" for sending messages and redo the logic in the "ChatForm" component; 2. Add a file with auxiliary functions; 3. Add a file with constants.
