+++ src/data/consts.js
@@ -1,0 +1,1 @@
+export const BOT_NAME = 'bot';
