+++ src/helper/helper.js
@@ -1,0 +1,7 @@
+export const sendMessage = (author, text, chatId) => {
+    const somebodyMessage = {
+        message: {author: author, text: text},
+        chatId: chatId,
+    };
+    return somebodyMessage
+};
