+++ src/store/ChatList/Action.js
@@ -1,3 +1,6 @@
+import { sendMessage } from '../../helper/helper';
+import { BOT_NAME } from '../../data/consts';
+
 export const ADD_MESSAGE_IN_CHAT_LIST = 'ADD_MESSAGE_IN_CHAT_LIST';
 
 export const addMessageInChatListAction = ({message, chatId}) => ({
@@ -14,3 +17,12 @@ export const removeMessageInChatListAction = (chatId) => ({
     type: REMOVE_MESSAGE_IN_CHAT_LIST,
     payload: chatId,
 });
+
+export const addMessageInChatListWithThunkAction = (author, text, chatId) => (dispatch) => {
+    const userMessage = sendMessage(author, text, chatId);
+    dispatch(addMessageInChatListAction(userMessage));
+    if (author !== BOT_NAME) {
+        const botMessage = sendMessage(BOT_NAME, `Ok, ${author}, принято!`, chatId);
+        dispatch(addMessageInChatListAction(botMessage));
+    }
+};
